requirements.txt
flask
pandas
scikit-learn
python-dotenv
flask-swagger-ui
flask-sqlalchemy
flask-migrate
